requirements.txt
Django==4.3
djangorestframework==3.21.0
djangorestframework-simplejwt==6.1.0
